+++ lib/internals/src/utils/common/is-not-nil.ts
@@ -3,3 +3,7 @@ import { Any } from '@ngxs-labs/data/typings';
 export function isNotNil(val: Any): boolean {
     return val !== 'undefined' && typeof val !== 'undefined' && val !== null;
 }
+
+export function isNullOrUndefined(val: Any): boolean {
+    return !isNotNil(val);
+}