+++ lib/typings/src/entity/entity-types.ts
@@ -1,0 +1,11 @@
+export type EntityIdType = string | number;
+
+export type EntityDictionary<K extends string | number, V> = {
+    [key in K]?: V;
+};
+
+export type EntityStateValue<T> = T | ((state: T) => T);
+
+export type EntityPatchValue<T> = Partial<T>;
+
+export type KeysDictionary<K extends string | number> = { [key in K]?: K };