+++ lib/typings/src/entity/ngxs-entity-collections.ts
@@ -1,0 +1,6 @@
+import { EntityDictionary, EntityIdType } from './entity-types';
+
+export interface NgxsEntityCollections<V, K extends number | string = EntityIdType> {
+    ids: K[];
+    entities: EntityDictionary<K, V>;
+}